chore: keyword Input 파일 분리
